+++ pyoxidizer/src/starlark/python_packaging_policy.rs
@@ -108,6 +108,7 @@ impl TypedValue for PythonPackagingPolicyValue {
 
     fn get_attr(&self, attribute: &str) -> ValueResult {
         let v = match attribute {
+            "allow_files" => Value::from(self.inner.allow_files()),
             "allow_in_memory_shared_library_loading" => {
                 Value::from(self.inner.allow_in_memory_shared_library_loading())
             }
@@ -149,6 +150,7 @@ impl TypedValue for PythonPackagingPolicyValue {
 
     fn has_attr(&self, attribute: &str) -> Result<bool, ValueError> {
         Ok(match attribute {
+            "allow_files" => true,
             "allow_in_memory_shared_library_loading" => true,
             "bytecode_optimize_level_zero" => true,
             "bytecode_optimize_level_one" => true,
@@ -167,6 +169,9 @@ impl TypedValue for PythonPackagingPolicyValue {
 
     fn set_attr(&mut self, attribute: &str, value: Value) -> Result<(), ValueError> {
         match attribute {
+            "allow_files" => {
+                self.inner.set_allow_files(value.to_bool());
+            }
             "allow_in_memory_shared_library_loading" => {
                 self.inner
                     .set_allow_in_memory_shared_library_loading(value.to_bool());
@@ -420,6 +425,14 @@ mod tests {
             "policy.resources_location_fallback = None; policy.resources_location_fallback",
         )?;
         assert_eq!(value.get_type(), "NoneType");
+
+        let value = env.eval("policy.allow_files")?;
+        assert_eq!(value.get_type(), "bool");
+        assert!(!value.to_bool());
+
+        let value = env.eval("policy.allow_files = True; policy.allow_files")?;
+        assert_eq!(value.get_type(), "bool");
+        assert!(value.to_bool());
 
         let value = env.eval("policy.allow_in_memory_shared_library_loading")?;
         assert_eq!(value.get_type(), "bool");